# src/config/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"SuccessCode": {
				"enum": [
					"OPERATION_SUCCESS"
				],
				"type": "string"
			},
			"GeneralResponse": {
				"properties": {
					"data": {},
					"code": {
						"$ref": "#/components/schemas/SuccessCode"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"data",
					"code",
					"message"
				],
				"type": "object"
			},
...
			"CreateBuildingDTO": {
				"properties": {
					"name": {
						"type": "string"
					},
					"areaId": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"areaId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateBuildingDTO": {
				"properties": {
					"name": {
						"type": "string"
					},
					"areaId": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
...
			"CreateAreaDTO": {
				"properties": {
					"name": {
						"type": "string"
					},
					"zoneId": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"zoneId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateAreaDTO": {
				"properties": {
					"name": {
						"type": "string"
					},
					"zoneId": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"_36_Enums.ApartmentStatus": {
				"type": "string",
				"enum": [
					"AVAILABLE",
					"RENTED",
					"UNAVAILABLE"
				]
			},
			"ApartmentStatus": {
				"$ref": "#/components/schemas/_36_Enums.ApartmentStatus"
			},
			"CreateApartmentImageDTO": {
				"properties": {
					"url": {
						"type": "string"
					}
				},
				"required": [
					"url"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateApartmentDTO": {
				"properties": {
					"description": {
						"type": "string"
					},
					"price": {
						"type": "number",
						"format": "double"
					},
					"areaSize": {
						"type": "number",
						"format": "double"
					},
					"bedrooms": {
						"type": "number",
						"format": "double"
					},
					"bathrooms": {
						"type": "number",
						"format": "double"
					},
					"priority": {
						"type": "number",
						"format": "double"
					},
					"status": {
						"$ref": "#/components/schemas/ApartmentStatus"
					},
					"contactUserId": {
						"type": "string"
					},
					"buildingId": {
						"type": "string"
					},
					"images": {
						"items": {
							"$ref": "#/components/schemas/CreateApartmentImageDTO"
						},
						"type": "array"
					}
				},
				"required": [
					"price",
					"areaSize",
					"bedrooms",
					"bathrooms",
					"contactUserId",
					"buildingId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateApartmentDTO": {
				"properties": {
					"description": {
						"type": "string"
					},
					"price": {
						"type": "number",
						"format": "double"
					},
					"areaSize": {
						"type": "number",
						"format": "double"
					},
					"bedrooms": {
						"type": "number",
						"format": "double"
					},
					"bathrooms": {
						"type": "number",
						"format": "double"
					},
					"priority": {
						"type": "number",
						"format": "double"
					},
					"status": {
						"$ref": "#/components/schemas/ApartmentStatus"
					},
					"contactUserId": {
						"type": "string"
					},
					"buildingId": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "apartment-api",
		"version": "1.0.1",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
...
		"/buildings": {
			"post": {
				"operationId": "CreateBuilding",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GeneralResponse"
								}
							}
						}
					}
				},
				"description": "Tạo Tòa nhà (Building) mới",
				"tags": [
					"Building Management"
				],
				"security": [
					{
						"jwt": [
							"ADMIN"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateBuildingDTO"
							}
						}
					}
				}
			}
		},
		"/buildings/{buildingId}": {
			"put": {
				"operationId": "UpdateBuilding",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GeneralResponse"
								}
							}
						}
					}
				},
				"description": "Cập nhật Tòa nhà",
				"tags": [
					"Building Management"
				],
				"security": [
					{
						"jwt": [
							"ADMIN"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "buildingId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateBuildingDTO"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteBuilding",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GeneralResponse"
								}
							}
						}
					}
				},
				"description": "Xóa mềm Tòa nhà",
				"tags": [
					"Building Management"
				],
				"security": [
					{
						"jwt": [
							"ADMIN"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "buildingId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"get": {
				"operationId": "GetBuildingById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GeneralResponse"
								}
							}
						}
					}
				},
				"description": "Lấy Tòa nhà bằng ID",
				"tags": [
					"Building Management"
				],
				"security": [
					{
						"jwt": [
							"ADMIN"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "buildingId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/buildings/by-area/{areaId}": {
			"get": {
				"operationId": "GetAllBuildingsByArea",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GeneralResponse"
								}
							}
						}
					}
				},
				"description": "Lấy tất cả Tòa nhà theo Khu vực (Area)",
				"tags": [
					"Building Management"
				],
				"security": [
					{
						"jwt": [
							"ADMIN"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "areaId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "pageIndex",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "pageSize",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/areas": {
			"post": {
				"operationId": "CreateArea",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GeneralResponse"
								}
							}
						}
					}
				},
				"description": "Tạo Khu vực (Area) mới",
				"tags": [
					"Area Management"
				],
				"security": [
					{
						"jwt": [
							"ADMIN"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateAreaDTO"
							}
						}
					}
				}
			}
		},
		"/areas/{areaId}": {
			"put": {
				"operationId": "UpdateArea",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GeneralResponse"
								}
							}
						}
					}
				},
				"description": "Cập nhật Khu vực",
				"tags": [
					"Area Management"
				],
				"security": [
					{
						"jwt": [
							"ADMIN"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "areaId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateAreaDTO"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteArea",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GeneralResponse"
								}
							}
						}
					}
				},
				"description": "Xóa mềm Khu vực",
				"tags": [
					"Area Management"
				],
				"security": [
					{
						"jwt": [
							"ADMIN"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "areaId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"get": {
				"operationId": "GetAreaById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GeneralResponse"
								}
							}
						}
					}
				},
				"description": "Lấy Khu vực bằng ID",
				"tags": [
					"Area Management"
				],
				"security": [
					{
						"jwt": [
							"ADMIN"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "areaId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/areas/by-zone/{zoneId}": {
			"get": {
				"operationId": "GetAllAreasByZone",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GeneralResponse"
								}
							}
						}
					}
				},
				"description": "Lấy tất cả Khu vực theo Phân khu (Zone)",
				"tags": [
					"Area Management"
				],
				"security": [
					{
						"jwt": [
							"ADMIN"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "zoneId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "pageIndex",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "pageSize",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/apartments": {
			"post": {
				"operationId": "CreateApartment",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GeneralResponse"
								}
							}
						}
					}
				},
				"description": "Tạo Căn hộ (Apartment) mới",
				"tags": [
					"Apartment Management"
				],
				"security": [
					{
						"jwt": [
							"ADMIN"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateApartmentDTO"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetAllApartments",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GeneralResponse"
								}
							}
						}
					}
				},
				"description": "Lấy tất cả Căn hộ (Public)",
				"tags": [
					"Apartment Management"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "pageIndex",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "pageSize",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/apartments/{apartmentId}": {
			"put": {
				"operationId": "UpdateApartment",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GeneralResponse"
								}
							}
						}
					}
				},
				"description": "Cập nhật Căn hộ",
				"tags": [
					"Apartment Management"
				],
				"security": [
					{
						"jwt": [
							"ADMIN"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "apartmentId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateApartmentDTO"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteApartment",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GeneralResponse"
								}
							}
						}
					}
				},
				"description": "Xóa mềm Căn hộ",
				"tags": [
					"Apartment Management"
				],
				"security": [
					{
						"jwt": [
							"ADMIN"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "apartmentId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"get": {
				"operationId": "GetApartmentById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GeneralResponse"
								}
							}
						}
					}
				},
				"description": "Lấy Căn hộ bằng ID (Public)",
				"tags": [
					"Apartment Management"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "apartmentId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/apartments/by-building/{buildingId}": {
			"get": {
				"operationId": "GetAllApartmentsByBuilding",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GeneralResponse"
								}
							}
						}
					}
				},
				"description": "Lấy tất cả Căn hộ theo Tòa nhà (Public)",
				"tags": [
					"Apartment Management"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "buildingId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "pageIndex",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "pageSize",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}